requirements.txt
aniso8601==7.0.0
appdirs==1.4.4
asgiref==3.2.10
atomicwrites==1.4.0
attrs==19.3.0
beautifulsoup4==4.9.1
blis==0.7.5
bootstrap4==0.1.0
catalogue==2.0.6
certifi==2021.10.8
chardet==3.0.4
charset-normalizer==2.0.8
click==7.1.2
colorama==0.4.4
conllu==4.4.1
coverage==5.2.1
cymem==2.0.6
defusedxml==0.6.0
diff-match-patch==20200713
distlib==0.3.1
Django==3.1
django-cors-headers==3.5.0
django-filter==2.3.0
django-import-export==2.3.0
django-mysql==3.8.1
django-rest-auth==0.9.5
django-rest-framework==0.1.0
django-six==1.0.4
djangorestframework==3.11.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
et-xmlfile==1.1.0
excel2json==1.0.1
Faker==4.1.2
filelock==3.0.12
geosky==1.0.9
googlemaps==4.5.3
graphene==2.1.8
graphene-django==2.13.0
graphql-core==2.3.2
graphql-relay==2.0.1
idna==3.3
importlib-metadata==4.8.2
inexactsearch==1.0.2
iniconfig==1.0.1
interchange==2021.0.4
jdcal==1.4.1
Jinja2==3.0.3
joblib==1.1.0
jsonfield==3.1.0
MarkupPy==1.14
MarkupSafe==2.0.1
monotonic==1.6
more-itertools==8.4.0
murmurhash==1.0.6
mysqlclient==2.0.1
negspacy==0.1.9
nltk==3.6.5
nmslib==2.1.1
odfpy==1.4.1
openpyxl==3.0.9
packaging==21.3
pandas==1.3.4
pansi==2020.7.3
pathy==0.6.1
Pillow==7.2.0
plac==1.1.3
pluggy==0.13.1
preshed==3.0.6
promise==2.3
psutil==5.8.0
py==1.9.0
py2neo==2021.2.3
pybind11==2.6.1
pydantic==1.7.4
Pygments==2.10.0
PyJWT==1.7.1
pymongo==4.0
pyparsing==3.0.6
pysbd==0.3.4
pyspellchecker==0.6.2
pytest==6.0.1
pytest-cov==2.10.0
pytest-django==3.9.0
python-dateutil==2.8.2
pytz==2021.3
PyYAML==5.3.1
regex==2021.11.10
requests==2.26.0
Rx==1.6.1
scikit-learn==1.0.1
scipy==1.7.3
scispacy==0.4.0
selenium==3.141.0
silpa-common==0.3
singledispatch==3.4.0.3
six==1.16.0
smart-open==5.2.1
soundex==1.1.3
soupsieve==2.0.1
spacy==3.0.5
spacy-legacy==3.0.8
spellchecker==0.4
sqlparse==0.3.1
srsly==2.4.2
tablib==2.0.0
text-unidecode==1.3
thinc==8.0.13
threadpoolctl==3.0.0
toml==0.10.1
tqdm==4.62.3
typer==0.3.2
typing-extensions==3.10.0.2
Unidecode==1.1.1
urllib3==1.26.7
utils==1.0.1
virtualenv==20.4.2
wasabi==0.8.2
wincertstore==0.2
xlrd==1.2.0
xlwt==1.3.0
zipp==3.6.0
